# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Wavelet 平台后端 API，提供用户认证、系统配置、任务调度等通用功能。",
        "title": "Wavelet API",
        "contact": {
            "name": "Wavelet",
            "url": "https://github.com/Rain-kl/Wavelet"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/admin/uploads/files": {
            "get": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "分页获取系统上传的文件列表，支持文件名关键词、业务类型、扩展名、上传用户ID过滤",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "获取文件列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码（默认 1）",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量（默认 20，最大 100）",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "文件名关键词（模糊匹配）",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "业务分类过滤",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "扩展名过滤",
                        "name": "extension",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "上传用户 ID 过滤",
                        "name": "user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Any"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handler.listFilesResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Any"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/uploads/files/{id}": {
            "delete": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "将指定 ID 的文件状态置为 deleted（软删除）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "删除文件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文件 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/response.Any"
                        }
                    },
                    "404": {
                        "description": "文件不存在",
                        "schema": {
                            "$ref": "#/definitions/response.Any"
                        }
                    }
                }
            }
        },
...
        "/api/v1/admin/uploads/types": {
            "get": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "查询系统内所有不重复的上传业务分类标识（如 avatar, doc 等）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "获取业务分类列表",
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Any"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/api/v1/upload/my": {
            "get": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "分页获取当前登录用户上传的文件，支持文件名关键词、业务类型、扩展名过滤",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "upload"
                ],
                "summary": "获取我的文件列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码（默认 1）",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量（默认 20，最大 100）",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "文件名关键词（模糊匹配）",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "业务分类过滤",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "扩展名过滤",
                        "name": "extension",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Any"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handler.listMyFilesResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/response.Any"
                        }
                    }
                }
            }
        },
        "/api/v1/upload/{id}": {
            "put": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "更新当前用户本人的文件名或访问权限模式 (AccessMode)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "upload"
                ],
                "summary": "更新我的文件信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文件 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新字段",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.updateMyFileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Any"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Upload"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "403": {
                        "description": "无权操作",
                        "schema": {
                            "$ref": "#/definitions/response.Any"
                        }
                    },
                    "404": {
                        "description": "文件不存在",
                        "schema": {
                            "$ref": "#/definitions/response.Any"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "将当前用户本人的文件状态置为 deleted（软删除）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "upload"
                ],
                "summary": "删除我的文件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文件 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/response.Any"
                        }
                    },
                    "403": {
                        "description": "无权操作",
                        "schema": {
                            "$ref": "#/definitions/response.Any"
                        }
                    },
                    "404": {
                        "description": "文件不存在",
                        "schema": {
                            "$ref": "#/definitions/response.Any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.listFilesResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Upload"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "handler.listMyFilesResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Upload"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "handler.updateMyFileRequest": {
            "type": "object",
            "properties": {
                "access_mode": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                },
                "file_name": {
                    "type": "string",
                    "maxLength": 255
                }
            }
        },
...
        "models.Upload": {
            "type": "object",
            "properties": {
                "access_mode": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "extension": {
                    "type": "string"
                },
                "file_name": {
                    "type": "string"
                },
                "file_path": {
                    "type": "string"
                },
                "file_size": {
                    "type": "integer"
                },
                "hash": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "0"
                },
                "metadata": {
                    "$ref": "#/definitions/models.UploadMetadata"
                },
                "mime_type": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.UploadStatus"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string",
                    "example": "0"
                }
            }
        },
        "models.UploadMetadata": {
            "type": "object",
            "properties": {
                "bucket": {
                    "type": "string"
                },
                "client_ip": {
                    "type": "string"
                },
                "duration": {
                    "type": "number"
                },
                "extra": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "height": {
                    "type": "integer"
                },
                "original_mime": {
                    "type": "string"
                },
                "user_agent": {
                    "type": "string"
                },
                "width": {
                    "type": "integer"
                }
            }
        },
        "models.UploadStatus": {
            "type": "string",
            "enum": [
                "pending",
                "used",
                "deleted"
            ],
            "x-enum-comments": {
                "UploadStatusDeleted": "已删除",
                "UploadStatusPending": "待使用",
                "UploadStatusUsed": "已使用"
            },
            "x-enum-descriptions": [
                "待使用",
                "已使用",
                "已删除"
            ],
            "x-enum-varnames": [
                "UploadStatusPending",
                "UploadStatusUsed",
                "UploadStatusDeleted"
            ]
        },
        "response.Any": {
            "type": "object",
            "properties": {
                "data": {},
                "error_msg": {
                    "type": "string",
                    "example": ""
                }
            }
        }
    },
    "securityDefinitions": {
        "SessionCookie": {
            "type": "apiKey",
            "name": "session",
            "in": "cookie"
        }
    }
}